I may have Isolated a pitch correction method if I'm very lucky, however I just need to pay attention to the multiplication of the index to get the frequency.

+++ JawJaw/src/pitchDetection/Mode.java
@@ -2,8 +2,8 @@ package pitchDetection;
 
 public class Mode {
 	
-	 public int getMode(int[] n){//method for calculating the mode of the pitch array generated
-	    int t = 0;
+	 public float getMode(float[] n){//method for calculating the mode of the pitch array generated
+	    float t = 0;
 	    for(int i=0; i<n.length; i++){
 	        for(int j=1; j<n.length-i; j++){
 	            if(n[j-1] > n[j]){
@@ -14,7 +14,7 @@ public class Mode {
 	        }
 	    }
 
-	    int mode = n[0];
+	    float mode = n[0];
 	    int temp = 1;
 	    int temp2 = 1;
 	    for(int i=1;i<n.length;i++){
